[YUNIKORN-2159] Clean up AppManager implementation
Refactoring of AppManager implementations to consolidate into AppManagementService.

- Merged general app manager into AppManagementService
- Migrated pkg/appmgmt/* -> pkg/cache/*
- Migrated pkg/callback/* -> pkg/cache/*
- Migrated various utils -> pkg/cache/ to avoid circular dependencies
- Removed Recoverable interface (functionality merged into
  AppManagementService)
- Removed AppManager interface (functionality merged into
  AppManagementService)
- Removed ManagedApp and ManagedTask interfaces
- Skipped (currently) broken recovery tests due to interfaces not being usable.
  These tests will be rewritten in a future PR once the recovery code is integrated
  into the context.
