+++ pkg/cache/appmgmt_recovery.go
@@ -16,7 +16,7 @@
  limitations under the License.
 */
 
-package appmgmt
+package cache
 
 import (
 	"errors"
@@ -25,9 +25,6 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/apache/yunikorn-k8shim/pkg/appmgmt/general"
-	"github.com/apache/yunikorn-k8shim/pkg/appmgmt/interfaces"
-	"github.com/apache/yunikorn-k8shim/pkg/cache"
 	"github.com/apache/yunikorn-k8shim/pkg/common/utils"
 	"github.com/apache/yunikorn-k8shim/pkg/log"
 )
@@ -46,47 +43,43 @@ func (svc *AppManagementService) WaitForRecovery() error {
 	return nil
 }
 
-func (svc *AppManagementService) recoverApps() (map[string]interfaces.ManagedApp, error) {
+func (svc *AppManagementService) recoverApps() (map[string]*Application, error) {
 	log.Log(log.ShimAppMgmt).Info("Starting app recovery")
-	recoveringApps := make(map[string]interfaces.ManagedApp)
-	for _, mgr := range svc.managers {
-		if m, ok := mgr.(interfaces.Recoverable); ok {
-			pods, err := m.ListPods()
-			if err != nil {
-				log.Log(log.ShimAppMgmt).Error("failed to list apps", zap.Error(err))
-				return recoveringApps, err
-			}
+	recoveringApps := make(map[string]*Application)
+	pods, err := svc.ListPods()
+	if err != nil {
+		log.Log(log.ShimAppMgmt).Error("failed to list apps", zap.Error(err))
+		return recoveringApps, err
+	}
 
-			sort.Slice(pods, func(i, j int) bool {
-				return pods[i].CreationTimestamp.Unix() < pods[j].CreationTimestamp.Unix()
-			})
+	sort.Slice(pods, func(i, j int) bool {
+		return pods[i].CreationTimestamp.Unix() < pods[j].CreationTimestamp.Unix()
+	})
 
-			// Track terminated pods that we have already seen in order to
-			// skip redundant handling of async events in RecoveryDone
-			// This filter is used for terminated pods to remain consistent
-			// with pod filters in the informer
-			terminatedYkPods := make(map[string]bool)
-			for _, pod := range pods {
-				if utils.GetApplicationIDFromPod(pod) != "" {
-					if !utils.IsPodTerminated(pod) {
-						app := svc.podEventHandler.HandleEvent(general.AddPod, general.Recovery, pod)
-						recoveringApps[app.GetApplicationID()] = app
-						continue
-					}
-					terminatedYkPods[string(pod.UID)] = true
-				}
+	// Track terminated pods that we have already seen in order to
+	// skip redundant handling of async events in RecoveryDone
+	// This filter is used for terminated pods to remain consistent
+	// with pod filters in the informer
+	terminatedYkPods := make(map[string]bool)
+	for _, pod := range pods {
+		if utils.GetApplicationIDFromPod(pod) != "" {
+			if !utils.IsPodTerminated(pod) {
+				app := svc.podEventHandler.HandleEvent(AddPod, Recovery, pod)
+				recoveringApps[app.GetApplicationID()] = app
+				continue
 			}
-			log.Log(log.ShimAppMgmt).Info("Recovery finished")
-			svc.podEventHandler.RecoveryDone(terminatedYkPods)
+			terminatedYkPods[string(pod.UID)] = true
 		}
 	}
+	log.Log(log.ShimAppMgmt).Info("Recovery finished")
+	svc.podEventHandler.RecoveryDone(terminatedYkPods)
 
 	return recoveringApps, nil
 }
 
 // waitForAppRecovery blocks until either all applications have been processed (returning true)
 // or cancelWaitForAppRecovery is called (returning false)
-func (svc *AppManagementService) waitForAppRecovery(recoveringApps map[string]interfaces.ManagedApp) bool {
+func (svc *AppManagementService) waitForAppRecovery(recoveringApps map[string]*Application) bool {
 	svc.cancelRecovery.Store(false) // reset cancellation token
 	recoveryStartTime := time.Now()
 	counter := 0
@@ -119,10 +112,10 @@ func (svc *AppManagementService) cancelWaitForAppRecovery() {
 }
 
 // removeRecoveredApps is used to walk the currently recovering apps list and remove those that have finished recovering
-func (svc *AppManagementService) removeRecoveredApps(recoveringApps map[string]interfaces.ManagedApp) {
+func (svc *AppManagementService) removeRecoveredApps(recoveringApps map[string]*Application) {
 	for _, app := range recoveringApps {
 		state := app.GetApplicationState()
-		if state != cache.ApplicationStates().New && state != cache.ApplicationStates().Recovering {
+		if state != ApplicationStates().New && state != ApplicationStates().Recovering {
 			log.Log(log.ShimAppMgmt).Info("Recovered application",
 				zap.String("appId", app.GetApplicationID()),
 				zap.String("state", state))
