+++ pkg/cache/task_sched_state_test.go
@@ -16,7 +16,7 @@
  limitations under the License.
 */
 
-package interfaces
+package cache
 
 import (
 	"testing"
