+++ pkg/conf/schedulerconf_test.go
@@ -122,7 +122,6 @@ func TestParseConfigMap(t *testing.T) {
 		{CMSvcVolumeBindTimeout, "VolumeBindTimeout", 15 * time.Second},
 		{CMSvcEventChannelCapacity, "EventChannelCapacity", 1234},
 		{CMSvcDispatchTimeout, "DispatchTimeout", 3 * time.Minute},
-		{CMSvcOperatorPlugins, "OperatorPlugins", "test-operators"},
 		{CMSvcDisableGangScheduling, "DisableGangScheduling", true},
 		{CMSvcEnableConfigHotRefresh, "EnableConfigHotRefresh", false},
 		{CMSvcPlaceholderImage, "PlaceHolderImage", "test-image"},
@@ -155,7 +154,6 @@ func TestUpdateConfigMapNonReloadable(t *testing.T) {
 		{CMSvcVolumeBindTimeout, "VolumeBindTimeout", 15 * time.Second, false},
 		{CMSvcEventChannelCapacity, "EventChannelCapacity", 1234, false},
 		{CMSvcDispatchTimeout, "DispatchTimeout", 3 * time.Minute, false},
-		{CMSvcOperatorPlugins, "OperatorPlugins", "test-operators", false},
 		{CMSvcDisableGangScheduling, "DisableGangScheduling", true, false},
 		{CMSvcPlaceholderImage, "PlaceHolderImage", "test-image", false},
 		{CMSvcNodeInstanceTypeNodeLabelKey, "InstanceTypeNodeLabelKey", "node.kubernetes.io/instance-type", false},
