+++ test/e2e/framework/helpers/k8s/gang_job.go
@@ -27,7 +27,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/apache/yunikorn-k8shim/pkg/appmgmt/interfaces"
+	"github.com/apache/yunikorn-k8shim/pkg/cache"
 	"github.com/apache/yunikorn-k8shim/pkg/common/constants"
 	"github.com/apache/yunikorn-k8shim/pkg/common/utils"
 )
@@ -59,7 +59,7 @@ func InitTestJob(jobName string, parallelism, completions int32, pod *v1.Pod) *b
 func getGangSchedulingAnnotations(placeholderTimeout int,
 	schedulingStyle string,
 	taskGroupName string,
-	taskGroups []*interfaces.TaskGroup) map[string]string {
+	taskGroups []*cache.TaskGroup) map[string]string {
 	annotations := make(map[string]string)
 	var schedulingParams string
 
@@ -93,7 +93,7 @@ func DecoratePodForGangScheduling(
 	placeholderTimeout int,
 	schedulingStyle string,
 	taskGroupName string,
-	taskGroups []*interfaces.TaskGroup,
+	taskGroups []*cache.TaskGroup,
 	pod *v1.Pod) *v1.Pod {
 	gangSchedulingAnnotations := getGangSchedulingAnnotations(placeholderTimeout, schedulingStyle, taskGroupName, taskGroups)
 	pod.Annotations = utils.MergeMaps(pod.Annotations, gangSchedulingAnnotations)
@@ -101,8 +101,8 @@ func DecoratePodForGangScheduling(
 	return pod
 }
 
-func InitTaskGroups(conf SleepPodConfig, mainTaskGroupName, secondTaskGroupName string, parallelism int) []*interfaces.TaskGroup {
-	tg1 := &interfaces.TaskGroup{
+func InitTaskGroups(conf SleepPodConfig, mainTaskGroupName, secondTaskGroupName string, parallelism int) []*cache.TaskGroup {
+	tg1 := &cache.TaskGroup{
 		MinMember: int32(parallelism),
 		Name:      mainTaskGroupName,
 		MinResource: map[string]resource.Quantity{
@@ -113,7 +113,7 @@ func InitTaskGroups(conf SleepPodConfig, mainTaskGroupName, secondTaskGroupName 
 
 	// create TG2 more with more members than needed, also make sure that
 	// placeholders will stay in Pending state
-	tg2 := &interfaces.TaskGroup{
+	tg2 := &cache.TaskGroup{
 		MinMember: int32(parallelism + 1),
 		Name:      secondTaskGroupName,
 		MinResource: map[string]resource.Quantity{
@@ -125,15 +125,15 @@ func InitTaskGroups(conf SleepPodConfig, mainTaskGroupName, secondTaskGroupName 
 		},
 	}
 
-	tGroups := make([]*interfaces.TaskGroup, 2)
+	tGroups := make([]*cache.TaskGroup, 2)
 	tGroups[0] = tg1
 	tGroups[1] = tg2
 
 	return tGroups
 }
 
-func InitTaskGroup(conf SleepPodConfig, taskGroupName string, parallelism int32) []*interfaces.TaskGroup {
-	tg1 := &interfaces.TaskGroup{
+func InitTaskGroup(conf SleepPodConfig, taskGroupName string, parallelism int32) []*cache.TaskGroup {
+	tg1 := &cache.TaskGroup{
 		MinMember: parallelism,
 		Name:      taskGroupName,
 		MinResource: map[string]resource.Quantity{
@@ -142,7 +142,7 @@ func InitTaskGroup(conf SleepPodConfig, taskGroupName string, parallelism int32)
 		},
 	}
 
-	tGroups := make([]*interfaces.TaskGroup, 1)
+	tGroups := make([]*cache.TaskGroup, 1)
 	tGroups[0] = tg1
 
 	return tGroups
